argus_mvp_online/requirements.txt
streamlit
torch
ultralytics
Pillow
opencv-python
numpy
